Refactors file readers
