+++ lib/reader/flow_reader.ex
@@ -2,9 +2,8 @@ defmodule Reader.FlowReader do
   alias Reader.Parser
   alias Repository.Record
 
-  def import(file_name) do
-    {:ok, record} = Record.start_link([])
-    flow = build_flow(file_name, record)
+  def import(stream, record) do
+    flow = build(stream, record)
 
     case Flow.run(flow) do
       :ok -> {:ok, record}
@@ -12,35 +11,12 @@ defmodule Reader.FlowReader do
     end
   end
 
-  defp build_flow(file_name, record) do
-    file_name
-    |> File.stream!()
-    |> Stream.map(&String.trim/1)
-    |> Stream.filter(&reject_empty_line/1)
-    |> Stream.chunk_while(
-      [],
-      fn element, acc ->
-        if Reader.Parser.check_default_format(element) do
-          case acc do
-            [] -> {:cont, [element]}
-            _ -> {:cont, acc |> Enum.reverse() |> Enum.join(" "), [element]}
-          end
-        else
-          {:cont, [element | acc]}
-        end
-      end,
-      fn
-        [] -> {:cont, []}
-        acc -> {:cont, acc |> Enum.reverse() |> Enum.join(" "), []}
-      end
-    )
+  defp build(stream, record) do
+    stream
     |> Flow.from_enumerable()
     |> Flow.partition()
     |> Flow.map(&store_message(record, Parser.parse_message(&1)))
   end
-
-  defp reject_empty_line(""), do: false
-  defp reject_empty_line(_line), do: true
 
   defp store_message(record, message) do
     %Message{author: author} = message
